pyproject.toml
[tool.poetry]
name = "bazartech-back"
version = "0.1.0"
description = ""
authors = ["Cácio Almeida <cacio.almeida@instruct.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.9"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
python-decouple = "^3.6"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
django-extensions = "^3.1.5"
dj-database-url = "^0.5.0"
djangorestframework-simplejwt = "^5.1.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.11.0"
whitenoise = "^6.0.0"
black = "^22.3.0"

[tool.poetry.dev-dependencies]
Werkzeug = "^2.1.2"
django-extensions = "^3.1.5"
isort = "^5.10.1"
black = "^22.3.0"
flake8 = "^4.0.1"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
line-length = 119
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

